Added function to perform dijkstras
Extended the tool library with a dijkstras function to find the
distance of the shortest path between two nodes in a connected graph.

Also made a minor improvement to the function that parses a file line
by line to not include any empty lines, along with adding a gitignore
file.