Add solution to day 5
This is the refined version with correct range splitting. In the
morning I brute-forced part two in Go instead as it would do the job
faster than my tired brain could figure out the smart solution.

Also added a read_with_delimeter function to be able to split the
input by a given delimiter. FOr this day that was a double new-line.