Add day 17 solution & remove dijkstras from tools
Finally got the shortest path-problem, but it was rather specific in
it's rules so it was difficult to represent the problem in a way to use
my previously built dijkstras.

The solution runs pretty slow on the input, but I am not sure of a way
to improve it immediately without missing out on requirements.