Add solution to day 20.
Finally solved all the problems for this year. This last problem was
hard for me to see through, but once started looking at the periodicity
of the signals it quickly dawned on me that I needed to just look at
the last step and check when all those inputs would go off at the same
time.